docs: fix up custom render + rerender samples (#226)
The get/query methods in the "how to override data-testid" sample
were not `.bind`ing correctly; I've replaced them with arrow functions
as suggested.  They were also passing in `container`, which wasn't
correct. Also, a `rerender` implementation is needed in such a
custom renderer, so I've added one to the sample.

I've also fixed up the separate custom rerender sample as it a)
incorrectly advertised an options parameter, and b) didn't pass
along the original baseElement.

For: kentcdodds/react-testing-library#225